UserViewModel refactoring
