+++ designer-website/Models/ViewModels/LoginViewModel.cs
@@ -1,0 +1,15 @@
+﻿using System.ComponentModel;
+using System.ComponentModel.DataAnnotations;
+using designer_website.Attributes;
+
+namespace designer_website.Models
+{
+    public class LoginViewModel : UserViewModel
+    {
+        [Required(ErrorMessage = "Это обязательное поле.")]
+        [EmailAddress(ErrorMessage = "Email адрес введен неправильно.")]
+        [DisplayName("Email address")]
+        [CheckEmailAvailability(true)]
+        public new string Email { get; set; }
+    }
+}
